+++ .travis.yml
@@ -1,9 +1,8 @@
 language: python
 python:
-  - "3.3"
-  - "3.4"
   - "3.5"
   - "3.6"
+  - "3.7"
 # command to install dependencies
 install: "pip install -r requirements.txt"
 # command to run tests
