+++ default.cfg
@@ -1,0 +1,7 @@
+[Default Configuration]
+DEBUG = True
+HOST = "127.0.0.1"
+PORT = 5000
+SECRET_KEY = "update_me"
+SQLALCHEMY_TRACK_MODIFICATIONS = False
+SQLALCHEMY_DATABASE_URI = "sqlite:///users.db"
